--- src/components/TopBasicNav.jsx
@@ -1,10 +1,0 @@
-import "./TopBasicNav.css";
-
-export default function TopBasicNav() {
-    return (
-        <nav className="topBasicNav">
-            <button className="prevBtn"></button>
-            <button className="moreBtn"></button>
-        </nav>
-    );
-}
