+++ src/components/loginjoin/LoginButton.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import "./LoginButton.css";
+
+function LoginButton(props) {
+    return (
+        <>
+            <input type="submit" id="login" value={props.value} />
+        </>
+    );
+}
+export default LoginButton;
