+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import TapMenu from "../components/TapMenu";
-import "./HomePage.css";
+import "./Home.css";
 
-export default function HomePage() {
+export default function Home() {
     return (
         <>
             <header className="topMainNav">
