+++ src/pages/Post.jsx
