+++ src/pages/style/UploadProduct.css
@@ -1,12 +1,3 @@
-.ir {
-    position: absolute;
-    left: -10000px;
-    top: auto;
-    width: 1px;
-    height: 1px;
-    overflow: hidden;
-}
-
 .saveBtn {
     width: 90px;
     background-color: var(--main-bg-color);
